% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_paseLista.R
\name{actualizar_pase_lista}
\alias{actualizar_pase_lista}
\title{Actualizar Estructura del Pase de Lista}
\usage{
actualizar_pase_lista(
  pool,
  id_proyecto,
  id_pase_lista,
  ids_encuestas_dialogo,
  id_cargo_supervisor = 37,
  dir_backup = "backups_pl"
)
}
\arguments{
\item{pool}{Conexión activa a la base de datos (DBI).}

\item{id_proyecto}{Numeric. ID del proyecto.}

\item{id_pase_lista}{Numeric. ID del cuestionario de pase de lista a modificar.}

\item{ids_encuestas_dialogo}{Numeric vector. IDs de los cuestionarios operativos.}

\item{id_cargo_supervisor}{Numeric. ID del cargo que funge como coordinador. Por defecto 37.}

\item{dir_backup}{Character. Carpeta donde se guardarán los respaldos del JSON. Por defecto "backups_pl".}
}
\value{
Invisible TRUE si la actualización en BD es exitosa.
}
\description{
Orquesta el proceso de reconstrucción dinámica del JSON de SurveyJS para el pase de lista.
Incluye un mecanismo de Backup automático del JSON original antes de cualquier modificación
para garantizar la recuperación ante desastres (ISO 27000).
}
