% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_pdf.R
\name{calcular_brigadistas_corte}
\alias{calcular_brigadistas_corte}
\title{Calcular brigadistas activos en una fecha de corte}
\usage{
calcular_brigadistas_corte(bd, corte, fill = "#6B3A8C", zonas = zonas)
}
\arguments{
\item{bd}{Un data frame que contiene los registros de brigadistas con las columnas \code{fecha}, \code{zona}, y \code{usuario_num}.}

\item{corte}{Una fecha específica (\code{Date}) para la cual se desea calcular el número de brigadistas activos por zona.}

\item{fill}{Un color hexadecimal para el relleno, utilizado en visualizaciones (por defecto es "#6B3A8C").}

\item{zonas}{Un vector de todas las zonas posibles para completar el conteo de brigadistas en zonas sin actividad en la fecha de corte.}
}
\value{
Un data frame con el número de brigadistas activos en cada zona en la fecha de corte, completando zonas faltantes y aplicando el color de relleno.
}
\description{
Calcula el número de brigadistas activos en cada zona en una fecha de corte específica, completando las zonas faltantes y aplicando un color de relleno.
}
\examples{
calcular_brigadistas_corte(bd, corte = as.Date("2024-12-31"), zonas = zonas)
}
