% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_pdf.R
\name{calcular_brigadistas_requeridos}
\alias{calcular_brigadistas_requeridos}
\title{Calcular número de brigadistas requeridos}
\usage{
calcular_brigadistas_requeridos(
  bd,
  dias_a_fin = dias_a_fin,
  semaforo = semaforo
)
}
\arguments{
\item{bd}{Un data frame con el progreso acumulado de diálogos efectivos, que incluye una columna \code{acumulado} que representa el total acumulado de diálogos hasta cada día.}

\item{dias_a_fin}{Número total de días restantes para alcanzar la meta final.}

\item{semaforo}{Un data frame que contiene la meta diaria de diálogos efectivos por zona, incluyendo una columna \code{meta_diaria} con la meta diaria de cada zona.}
}
\value{
Un data frame con el cálculo diario del número de brigadistas requeridos para alcanzar la meta de diálogos, basado en un promedio de 15 diálogos por brigadista por día.
}
\description{
Calcula el número de brigadistas requeridos diariamente para alcanzar la meta de diálogos efectivos antes de la fecha de fin.
}
\examples{
calcular_brigadistas_requeridos(bd, dias_a_fin = 20, semaforo = semaforo)
}
