% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_pdf.R
\name{calcular_conocimiento}
\alias{calcular_conocimiento}
\title{Calcular conocimiento en diálogos efectivos}
\usage{
calcular_conocimiento(bd, ...)
}
\arguments{
\item{bd}{Un data frame con los registros de diálogos, que incluye la columna \code{desglose} para filtrar los diálogos efectivos.}

\item{...}{Variables adicionales para agrupar y contar, permitiendo segmentar el conocimiento en distintas categorías.}
}
\value{
Un data frame con el conteo de diálogos efectivos por cada categoría especificada y su proporción (\code{media}).
}
\description{
Calcula el número de diálogos efectivos por categorías especificadas y su proporción, permitiendo el análisis de distribución de conocimiento en el equipo.
}
\examples{
calcular_conocimiento(bd, zona, usuario_num)
}
