% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_pdf.R
\name{calcular_dias_fin}
\alias{calcular_dias_fin}
\title{Calcular días hasta fecha de fin excluyendo festivos y domingos}
\usage{
calcular_dias_fin(bd, fecha_fin, dias_festivos)
}
\arguments{
\item{bd}{Un data frame que contiene una columna \code{fecha} con las fechas iniciales.}

\item{fecha_fin}{Una fecha de fin (\code{Date}) que indica el límite superior del rango de fechas a calcular.}
}
\value{
Un número entero que representa el total de días hábiles (excluyendo domingos y días festivos) entre la primera fecha de \code{bd} y \code{fecha_fin}.
}
\description{
Calcula el número de días entre la primera fecha en los datos y una fecha de fin específica, excluyendo días festivos y domingos.
}
\examples{
calcular_dias_fin(bd, fecha_fin = as.Date("2024-12-31"))
}
