% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_pdf.R
\name{calcular_dias_trabajados}
\alias{calcular_dias_trabajados}
\title{Calcular días trabajados}
\usage{
calcular_dias_trabajados(bd, dias_festivos = NULL)
}
\arguments{
\item{bd}{Un data frame que contiene una columna \code{fecha} con las fechas de los días trabajados.}

\item{dias_festivos}{Un vector de fechas correspondientes a los días festivos.}
}
\value{
Un número entero que representa el total de días trabajados.
}
\description{
Calcula el número de días trabajados excluyendo fines de semana y días festivos.
}
\examples{
calcular_dias_trabajados(bd, dias_festivos = c("2024-01-01", "2024-12-25"))
}
