% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_pdf.R
\name{calcular_horas_trabajadas}
\alias{calcular_horas_trabajadas}
\title{Calcular promedio de horas trabajadas por brigadista hasta la fecha de corte}
\usage{
calcular_horas_trabajadas(bd, corte)
}
\arguments{
\item{bd}{Un data frame con registros de inicio y fin de cada jornada, que incluye las columnas \code{fecha}, \code{usuario_num}, y \code{fecha_inicio} (tiempo de inicio de la jornada).}

\item{corte}{Una fecha de corte (\code{Date}) hasta la cual se desea completar la serie de fechas diarias.}
}
\value{
Un data frame con el promedio diario de horas trabajadas por brigadista, completando días faltantes hasta la fecha de corte y asignando un valor de 0 en días sin registros.
}
\description{
Calcula el promedio de horas trabajadas por brigadista en cada día, excluyendo jornadas de más de 16 horas, y completa las fechas faltantes hasta la fecha de corte.
}
\examples{
calcular_horas_trabajadas(bd, corte = as.Date("2024-12-31"))
}
