% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_pdf.R
\name{calcular_meta_zona}
\alias{calcular_meta_zona}
\title{Calcular meta por zona}
\usage{
calcular_meta_zona(zona_seccion, metas_secc, p = 0.7, unidad)
}
\arguments{
\item{zona_seccion}{Un data frame que contiene las secciones de la zona.}

\item{p}{Un número (opcional, por defecto 0.7) que indica el porcentaje de viviendas a considerar como meta.}

\item{unidad}{La variable de agrupación utilizada para calcular la meta en cada grupo.}

\item{entidad}{Una cadena de texto que indica la entidad de interés.}
}
\value{
Un data frame con la meta de viviendas calculada por unidad.
}
\description{
Calcula la meta de viviendas por zona para una entidad específica, utilizando datos del censo y un parámetro de ajuste.
}
\examples{
calcular_meta_zona("EntidadA", zona_seccion, p = 0.8, unidad = grupo)
}
