% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_pdf.R
\name{calcular_opinion}
\alias{calcular_opinion}
\title{Calcular distribución de opiniones en diálogos efectivos}
\usage{
calcular_opinion(
  bd,
  var_conocimiento,
  var_opinion,
  levels,
  labels,
  fill = "#6B3A8C",
  unidad = NULL
)
}
\arguments{
\item{bd}{Un data frame con los registros de diálogos, que incluye la columna \code{desglose} para filtrar los diálogos efectivos.}

\item{var_conocimiento}{Una cadena que indica el nombre de la columna que evalúa el conocimiento (donde el valor debe ser "Sí").}

\item{var_opinion}{Una cadena que indica el nombre de la columna que representa la variable de opinión a analizar.}

\item{levels}{Un vector con los niveles para ordenar la variable de opinión.}

\item{labels}{Un vector con las etiquetas para cada nivel de la variable de opinión.}

\item{fill}{Un color hexadecimal para el relleno, utilizado en visualizaciones (por defecto es "#6B3A8C").}

\item{unidad}{Una cadena opcional que indica el nombre de una columna para agrupar la opinión por una unidad específica (por defecto es \code{NULL}).}
}
\value{
Un data frame con el conteo y proporción de opiniones sobre la variable seleccionada, incluyendo una etiqueta de porcentaje y un color de relleno.
}
\description{
Calcula la distribución de opiniones sobre una variable específica en diálogos efectivos, permitiendo agrupar por una unidad opcional.
}
