% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_pdf.R
\name{calcular_semaforo}
\alias{calcular_semaforo}
\title{Calcular semáforo de diálogos efectivos por zona}
\usage{
calcular_semaforo(bd, meta_zona, dias_fin)
}
\arguments{
\item{bd}{Un data frame con los datos de diálogos, que debe incluir las columnas \code{desglose} y \code{zona}.}

\item{meta_zona}{Un data frame que contiene la meta de diálogos por zona, con columnas \code{zona} y \code{meta}.}

\item{dias_fin}{Número total de días hábiles disponibles para alcanzar la meta.}
}
\value{
Un data frame con el cálculo del semáforo por zona, incluyendo el porcentaje de cumplimiento (\code{semaforo}) y un color representativo (\code{color}).
}
\description{
Calcula el nivel de cumplimiento de los diálogos efectivos por zona, generando un semáforo de colores según el progreso respecto a la meta diaria.
}
\examples{
calcular_semaforo(bd, meta_zona, dias_fin = 20)
}
