% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_pdf.R
\name{calcular_tabla_zona}
\alias{calcular_tabla_zona}
\title{Calcular tabla resumen por zona}
\usage{
calcular_tabla_zona(bd = bd)
}
\value{
Un data frame que muestra un resumen detallado por zona, incluyendo información sobre visitas y avance de metas en formato porcentual y numérico.
}
\description{
Genera una tabla de resumen por zona que incluye la meta final, meta al día de corte, viviendas visitadas, diálogos efectivos y el porcentaje de avance respecto a la meta del día.
}
\examples{
calcular_tabla_zona()
}
