% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{conectar_base_datos}
\alias{conectar_base_datos}
\title{Conectar a la Base de Datos SQL Server (Pool)}
\usage{
conectar_base_datos()
}
\value{
Un objeto de clase \code{Pool} listo para ser utilizado con \code{dplyr} o \code{DBI}.
}
\description{
Establece una conexión segura a la base de datos SQL Server leyendo las credenciales
directamente desde las variables de entorno (\code{.Renviron}).
Detecta automáticamente el entorno de ejecución (Local vs. Positron Connect)
para asignar el driver ODBC correcto.
}
\details{
\strong{Selección Automática de Driver:}
\itemize{
\item Si existe la variable de entorno \code{pool_driver}, utiliza esa.
\item Si el Sistema Operativo es Windows o Mac (Darwin), utiliza \verb{ODBC Driver 17 for SQL Server}.
\item Si el Sistema Operativo es Linux (Positron Connect), utiliza \code{FreeTDS}.
}

\strong{Cumplimiento ISO 27001:} Previene la exposición de credenciales mediante el uso
estricto de variables de entorno.
}
