% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_pdf.R
\name{creacion_mapas_dialogos}
\alias{creacion_mapas_dialogos}
\title{Funcion creadora mapa de dialogos efectivos}
\usage{
creacion_mapas_dialogos(
  bd,
  nombre_mapa,
  label_mapa,
  color_low = "#9FA3BF",
  color_high = "#351062",
  color_label = "black",
  fill_label = "white",
  etiquetas = FALSE
)
}
\arguments{
\item{bd}{Un data frame que contiene una columna \code{n} con el numnero de observaciones o registrs}

\item{nombre_mapa}{Nombre que se le asigna al label del gradiente}

\item{label_mapa}{Toma una variable de \code{bd}  y lo muestra como los labels de los poligonos}

\item{color_low}{Color minimimo del gradiente}

\item{color_high}{Nombre maximo del gradiente}
}
\value{
Un mapa con degradado segun los valores de \code{n} en cada poligono
}
\description{
Grafica el el avance de dialogos, o de n elementos en un mapa
}
\examples{
calcular_dias_trabajados(bd, dias_festivos = c("2024-01-01", "2024-12-25"))
}
