% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_pdf.R
\name{crear_shp_candidato}
\alias{crear_shp_candidato}
\title{Crear capa espacial filtrada para un candidato}
\usage{
crear_shp_candidato(shp, meta, bd, variable, valor)
}
\arguments{
\item{shp}{Un objeto \code{sf} que representa la capa espacial base (shapefile).}

\item{meta}{Un data frame que contiene la meta de cada unidad espacial.}

\item{bd}{Un data frame que contiene datos de actividad, incluyendo la variable a filtrar.}

\item{variable}{Una cadena que indica el nombre de la variable en \code{bd} para aplicar el filtro.}

\item{valor}{El valor específico de \code{variable} que se desea filtrar en \code{bd}.}
}
\value{
Un objeto \code{sf} con la capa espacial combinada y filtrada según la proporción de actividad y el valor de la variable especificada.
}
\description{
Crea una capa espacial (\code{shp}) filtrada según una variable específica, al combinar datos de meta y actividad.
}
\examples{
crear_shp_candidato(shp, meta, bd, variable = "apoyo", valor = "Alto")
}
