% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_pdf.R
\name{crear_tabla_desglose}
\alias{crear_tabla_desglose}
\title{Crear tabla de desglose}
\usage{
crear_tabla_desglose(
  bd,
  bd_juarez = NULL,
  dias_festivos,
  promedio_efec_voc_diario = NA
)
}
\arguments{
\item{bd}{Un data frame que contiene datos de visitas y diálogos, incluyendo una columna \code{desglose}.}

\item{dias_festivos}{Un vector de fechas correspondientes a los días festivos que deben excluirse del cálculo de días trabajados.}
}
\value{
Un data frame con el desglose de los resultados, organizado por categorías específicas.
}
\description{
Genera una tabla con el desglose de los resultados de una campaña, incluyendo métricas como días trabajados y viviendas visitadas.
}
\examples{
crear_tabla_desglose(bd, dias_festivos = c("2024-01-01", "2024-12-25"))
}
