% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_pdf.R
\name{crear_tabla_dialogos}
\alias{crear_tabla_dialogos}
\title{Crear tabla de diálogos}
\usage{
crear_tabla_dialogos(bd, lower, upper, dias_festivos, bd_cruda = NULL)
}
\arguments{
\item{bd}{Un data frame que contiene datos de diálogos y columnas de interés como \code{desglose}, \code{duracion_minutos}, \code{numero}, \code{direccion_correo}, y \code{redes}.}

\item{lower}{Un número que define el límite inferior de la duración ideal de un diálogo en minutos.}

\item{upper}{Un número que define el límite superior de la duración ideal de un diálogo en minutos.}
}
\value{
Un data frame con el resumen de diálogos efectivos y otros indicadores.
}
\description{
Genera un resumen de los diálogos efectivos en una tabla, incluyendo métricas de efectividad y datos recabados.
}
\examples{
crear_tabla_dialogos(bd, lower = 5, upper = 15)
}
