% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_pdf.R
\name{crear_tabla_usuarios}
\alias{crear_tabla_usuarios}
\title{Genera una tabla resumen de usuarios y su actividad}
\usage{
crear_tabla_usuarios(bd, voceros_alta, coordinadores_alta, corte)
}
\arguments{
\item{bd}{Un data frame que contiene los datos de actividad de los usuarios. Debe incluir las columnas:
\itemize{
\item \code{fecha}: Fecha de la actividad.
\item \code{corte}: Fecha de corte de los datos.
\item \code{usuario_num}: Identificador único de usuario.
\item \code{fecha_inicio}: Hora de inicio de actividad del usuario.
\item \code{fecha_fin}: Hora de fin de actividad del usuario.
}}

\item{voceros_alta}{Número total de voceros dados de alta.}

\item{coordinadores_alta}{Número total de coordinadores dados de alta.}
}
\value{
Un data frame con las siguientes columnas:
\itemize{
\item \code{name}: Descripción de la métrica calculada.
\item \code{value}: Valor de la métrica calculada.
}
}
\description{
Esta función procesa una base de datos de actividad de usuarios y genera un resumen
con métricas clave como el número de usuarios activos, diálogos promedio por usuario,
horas trabajadas y porcentaje de usuarios con al menos 6 horas laboradas.
}
