% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_pdf.R
\name{graficar_efectivos_historico}
\alias{graficar_efectivos_historico}
\title{Graficar el historial de diálogos efectivos}
\usage{
graficar_efectivos_historico(
  bd,
  color_linea_acum,
  color_meta_me,
  color_zona = "#e9c46a",
  acumulado,
  interactivo = T,
  total_usuario,
  meta_diaria,
  corte = Sys.Date() - 1
)
}
\arguments{
\item{bd}{Un data frame con los registros de diálogos efectivos, incluyendo columnas como \code{fecha}, \code{acumulado}, y \code{n}.}

\item{color_linea_acum}{Color hexadecimal para la línea acumulada de diálogos.}

\item{color_meta_me}{Color hexadecimal para la línea de la meta mejor escenario.}

\item{color_zona}{Color hexadecimal para el área de la zona de referencia (por defecto es "#e9c46a").}

\item{acumulado}{Lógico. Si es \code{TRUE}, grafica el total acumulado; si es \code{FALSE}, grafica el conteo diario.}

\item{interactivo}{Lógico. Si es \code{TRUE}, genera una gráfica interactiva con \code{highcharter}; si es \code{FALSE}, usa \code{ggplot2}.}

\item{total_usuario}{Número total de usuarios o brigadistas activos, usado para calcular la meta diaria acumulada.}

\item{meta_diaria}{Meta diaria de diálogos efectivos, utilizada en el cálculo de la meta acumulada.}
}
\value{
Una gráfica de diálogos efectivos acumulados o diarios, ya sea interactiva o estática.
}
\description{
Genera una gráfica interactiva o estática del historial de diálogos efectivos acumulados o diarios, con la opción de agregar una meta diaria acumulada.
}
\examples{
graficar_efectivos_historico(bd, color_linea_acum = "#3498db", color_meta_me = "#e74c3c", acumulado = TRUE, interactivo = FALSE, total_usuario = 50, meta_diaria = 100)
}
