% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_aux.R
\name{obtener_brigadas}
\alias{obtener_brigadas}
\title{Obtiene la lista de brigadas de un proyecto}
\usage{
obtener_brigadas(pool, id_proyecto)
}
\arguments{
\item{pool}{Conexión a la base de datos.}

\item{id_proyecto}{Identificador del proyecto.}
}
\value{
Un tibble con la lista de brigadas, incluyendo id, distrito y nombre de brigada.
}
\description{
Obtiene la lista de brigadas de un proyecto
}
