% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_aux.R
\name{obtener_coordinadores}
\alias{obtener_coordinadores}
\title{Obtiene la lista de coordinadores de un proyecto}
\usage{
obtener_coordinadores(pool, id_proyecto, brigadas)
}
\arguments{
\item{pool}{Conexión a la base de datos.}

\item{id_proyecto}{Identificador del proyecto.}

\item{brigadas}{Tibble con la lista de brigadas.}
}
\value{
Un tibble con la lista de coordinadores, incluyendo id, municipio, nombre completo, cargo y otros detalles.
}
\description{
Obtiene la lista de coordinadores de un proyecto
}
