% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_aux.R
\name{obtener_voceros}
\alias{obtener_voceros}
\title{Obtiene la lista de voceros de un proyecto}
\usage{
obtener_voceros(pool, id_proyecto)
}
\arguments{
\item{pool}{Conexión a la base de datos.}

\item{id_proyecto}{Identificador del proyecto.}
}
\value{
Un tibble con la lista de voceros, incluyendo id, municipio, nombre completo, número y otros detalles.
}
\description{
Obtiene la lista de voceros de un proyecto
}
