% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_reporte_auditoria.R
\name{crear_workbook_auditoria}
\alias{crear_workbook_auditoria}
\title{Crear Libro de Excel para Reporte de Auditoría}
\usage{
crear_workbook_auditoria(datos_auditoria)
}
\arguments{
\item{datos_auditoria}{Una lista que debe contener obligatoriamente los elementos \code{res_auditoria},
\code{observaciones} y \code{res_auditoria_hist}.}
}
\value{
Un objeto de tipo \code{Workbook} (paquete \code{openxlsx}) listo para ser guardado con \code{saveWorkbook}.
}
\description{
Toma la lista de datos generada por \code{\link{generar_reporte_metricas}} y construye un objeto Workbook
de Excel con formato condicional de escala de colores (Rojo a Verde) para la columna "Promedio de evaluaciones".
}
\examples{
\dontrun{
wb <- crear_workbook_auditoria(reportes)
openxlsx::saveWorkbook(wb, "Reporte_Auditoria.xlsx", overwrite = TRUE)
}

}
